src/main/java/com/example/demotdd/ProductService.java
package com.example.demotdd;

import com.example.demotdd.entity.Product;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Service;
import org.springframework.web.server.ResponseStatusException;

@Service
@RequiredArgsConstructor
@Slf4j
public class ProductService {

    private final ProductRepository productRepository;

    public Product getBySku(String sku) {
        if ("i11pro".equals(sku)) {
            log.info("Esto busca a un Ijone 11");
        }
        return productRepository.getBySku(sku)
                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
    }
}
